# docs/openapi.yaml
openapi: 3.0.0
paths:
...
  /v1/me:
    get:
      operationId: ProfilesController_me
      parameters: []
      responses:
        '200':
          description: Current role-aware Meal Direct session.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeSessionEnvelopeDto'
        '401': &a3
          description: Missing, invalid, or expired Supabase JWT.
      security: &a4
        - supabaseAuth: []
      tags: &a5
        - me
    patch:
      operationId: ProfilesController_updateProfile
      parameters: []
      responses:
        '200':
          description: Updated safe profile fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProfileEnvelopeDto'
        '400':
          description: Invalid profile input.
        '401': *a3
      security: *a4
      tags: *a5
...
  /v1/referrals/me:
    get:
      operationId: ReferralsController_me
      parameters: []
      responses:
        '200':
          description: The current user referral code and how many users they referred.
        '401': &a6
          description: Missing, invalid, or expired Supabase JWT.
      security: &a7
        - supabaseAuth: []
      tags: &a8
        - referrals
  /v1/referrals/redeem:
    post:
      operationId: ReferralsController_redeem
      parameters: []
      responses:
        '200':
          description: Binds the current user to the referrer behind the code.
        '400':
          description: Invalid, unknown, or self-owned referral code.
        '401': *a6
        '409':
          description: A code was already redeemed, or onboarding is already complete.
      security: *a7
      tags: *a8
...
  /v1/admin/payments/{paymentId}:
    get:
      operationId: AdminPaymentsController_getPayment
      parameters:
        - name: paymentId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Admin-visible enriched payment detail scoped by campus for campus
            admins.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminPaymentDetailEnvelopeDto'
        '401': *a37
        '403': *a38
      security: *a39
      tags: *a40
...
  /v1/admin/payments/{paymentId}/notes:
    post:
      operationId: AdminPaymentsController_addNote
      parameters:
        - name: paymentId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Appends an internal admin note to the payment.
        '401': *a37
        '403': *a38
      security: *a39
      tags: *a40
...
  /v1/admin/refunds:
    get:
      operationId: RefundsController_listRefunds
      parameters: []
      responses:
        '200':
          description: Refund records scoped by campus for campus admins.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminRefundListEnvelopeDto'
        '401': &a53
          description: Missing, invalid, or expired Supabase JWT.
        '403': &a54
          description: Admin role is required.
      security: &a55
        - supabaseAuth: []
      tags: &a56
        - admin-refunds
  /v1/admin/refunds/{refundId}:
    get:
      operationId: RefundsController_getRefund
      parameters:
        - name: refundId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Refund detail.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminRefundEnvelopeDto'
        '401': *a53
        '403': *a54
      security: *a55
      tags: *a56
  /v1/admin/refunds/{refundId}/retry:
    post:
      operationId: RefundsController_retryRefund
      parameters:
        - name: refundId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Retries a failed refund against Paystack.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminRefundEnvelopeDto'
        '401': *a53
        '403': *a54
      security: *a55
      tags: *a56
  /v1/admin/refunds/{refundId}/resolve:
    post:
      operationId: RefundsController_resolveRefund
      parameters:
        - name: refundId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Records a manual refund resolution.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminRefundEnvelopeDto'
        '401': *a53
        '403': *a54
      security: *a55
      tags: *a56
  /v1/admin/refunds/{refundId}/initiate:
    post:
      operationId: RefundsController_initiateRefund
      parameters:
        - name: refundId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Pushes an approved refund to Paystack (money-moving step).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminRefundEnvelopeDto'
        '401': *a53
        '403': *a54
      security: *a55
      tags: *a56
  /v1/admin/refunds/{refundId}/mark-manually-resolved:
    post:
      operationId: RefundsController_markManuallyResolved
      parameters:
        - name: refundId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Marks a refund as manually resolved (settled out-of-band). Records
            the note and audits.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminRefundEnvelopeDto'
        '401': *a53
        '403': *a54
      security: *a55
      tags: *a56
  /v1/admin/refunds/{refundId}/notes:
    post:
      operationId: RefundsController_addNote
      parameters:
        - name: refundId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Appends an internal admin note to the refund.
        '401': *a53
        '403': *a54
      security: *a55
      tags: *a56
  /v1/admin/refunds/{refundId}/approve:
    post:
      operationId: RefundsController_approveRefund
      parameters:
        - name: refundId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Approves a requested refund.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminRefundEnvelopeDto'
        '401': *a53
        '403': *a54
      security: *a55
      tags: *a56
  /v1/admin/refunds/{refundId}/reject:
    post:
      operationId: RefundsController_rejectRefund
      parameters:
        - name: refundId
          required: true
          in: path
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: Rejects a requested refund.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AdminRefundEnvelopeDto'
        '401': *a53
        '403': *a54
      security: *a55
      tags: *a56
...
info:
  title: Meal Direct API
  description: Stable REST API for Meal Direct web and future mobile clients.
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    supabaseAuth:
      scheme: bearer
      bearerFormat: Supabase JWT
      type: http
      description: Supabase Auth access token.
    idempotencyKey:
      type: apiKey
      in: header
      name: Idempotency-Key
      description: Required for important mutations such as order creation and payments.
    requestId:
      type: apiKey
      in: header
      name: x-request-id
      description: Optional caller-provided request ID. The API echoes or generates one.
    operationsToken:
      scheme: bearer
      bearerFormat: JWT
      type: http
      description: Internal operations token. Temporary until Super Admin JWT/RBAC is
        implemented.
  schemas:
...
    MeActorDto:
      type: object
      properties:
        userId:
          type: string
          format: uuid
        role:
          type: string
          enum: &a94
            - customer
            - vendor
            - rider
            - campus_admin
            - super_admin
        email:
          type: string
        campusId:
          type: string
          format: uuid
        vendorId:
          type: string
          format: uuid
        riderId:
          type: string
          format: uuid
      required:
        - userId
        - role
    ProfileRecordDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          nullable: true
        displayName:
          type: string
          nullable: true
        phoneNumber:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        accountStatus:
          type: string
          enum:
            - active
            - suspended
            - deactivated
        defaultCampusId:
          type: string
          format: uuid
          nullable: true
        defaultLocationId:
          type: string
          format: uuid
          nullable: true
        onboardingCompletedAt:
          type: string
          nullable: true
        onboardingCompleted:
          type: boolean
        lastSeenAt:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - accountStatus
        - onboardingCompleted
        - createdAt
        - updatedAt
    CampusMembershipDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        campusId:
          type: string
          format: uuid
        campusName:
          type: string
        campusSlug:
          type: string
        timezone:
          type: string
        currency:
          type: string
        countryCode:
          type: string
        active:
          type: boolean
        joinedAt:
          type: string
      required:
        - id
        - campusId
        - campusName
        - campusSlug
        - timezone
        - currency
        - countryCode
        - active
        - joinedAt
    VendorMembershipDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        vendorId:
          type: string
          format: uuid
        campusId:
          type: string
          format: uuid
        vendorDisplayName:
          type: string
        vendorSlug:
          type: string
        role:
          type: string
          enum:
            - owner
            - staff
        active:
          type: boolean
      required:
        - id
        - vendorId
        - campusId
        - vendorDisplayName
        - vendorSlug
        - role
        - active
    RiderProfileDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        campusId:
          type: string
          format: uuid
        displayName:
          type: string
        phone:
          type: string
        status:
          type: string
          enum:
            - pending
            - verified
            - suspended
            - deactivated
        active:
          type: boolean
        available:
          type: boolean
      required:
        - id
        - campusId
        - displayName
        - phone
        - status
        - active
        - available
    AdminMembershipDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        campusId:
          type: string
          format: uuid
          nullable: true
        campusName:
          type: string
          nullable: true
        role:
          type: string
          enum:
            - campus_admin
            - super_admin
        active:
          type: boolean
        grantedAt:
          type: string
      required:
        - id
        - role
        - active
        - grantedAt
    MeSessionDto:
      type: object
      properties:
        actor:
          $ref: '#/components/schemas/MeActorDto'
        profile:
          $ref: '#/components/schemas/ProfileRecordDto'
        roles:
          type: array
          items:
            type: string
            enum: *a94
        campuses:
          type: array
          items:
            $ref: '#/components/schemas/CampusMembershipDto'
        vendorMemberships:
          type: array
          items:
            $ref: '#/components/schemas/VendorMembershipDto'
        riderProfiles:
          type: array
          items:
            $ref: '#/components/schemas/RiderProfileDto'
        adminMemberships:
          type: array
          items:
            $ref: '#/components/schemas/AdminMembershipDto'
      required:
        - actor
        - profile
        - roles
        - campuses
        - vendorMemberships
        - riderProfiles
        - adminMemberships
    MeSessionEnvelopeDto:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/MeSessionDto'
      required:
        - data
...
    ProfileEnvelopeDto:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/ProfileRecordDto'
      required:
        - data
...
    AdminPaymentDetailDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
        orderNumber:
          type: string
        customerId:
          type: string
          format: uuid
        customerEmail:
          type: string
          nullable: true
        campusId:
          type: string
          format: uuid
        orderStatus:
          type: string
        orderTotalKobo:
          type: number
        providerReference:
          type: string
        paymentStatus:
          type: string
        expectedAmountKobo:
          type: number
        paidAmountKobo:
          type: number
          nullable: true
        providerTransactionId:
          type: string
          nullable: true
        currency:
          type: string
        initializedAt:
          type: string
        paidAt:
          type: string
          nullable: true
        verifiedAt:
          type: string
          nullable: true
        webhookReceived:
          type: boolean
        webhookCount:
          type: number
        verificationStatus:
          type: string
          enum:
            - verified
            - unverified
        refundStatus:
          type: string
        refundedAmountKobo:
          type: number
        settlementImpactKobo:
          type: number
      required:
        - id
        - orderId
        - orderNumber
        - customerId
        - campusId
        - orderStatus
        - orderTotalKobo
        - providerReference
        - paymentStatus
        - expectedAmountKobo
        - currency
        - initializedAt
        - webhookReceived
        - webhookCount
        - verificationStatus
        - refundStatus
        - refundedAmountKobo
        - settlementImpactKobo
    AdminPaymentDetailEnvelopeDto:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/AdminPaymentDetailDto'
      required:
        - data
...
    AdminRefundRecordDto:
      type: object
      properties:
        id:
          type: string
          format: uuid
        paymentId:
          type: string
          format: uuid
        orderId:
          type: string
          format: uuid
        orderNumber:
          type: string
          nullable: true
        campusId:
          type: string
          nullable: true
        vendorId:
          type: string
          nullable: true
        customerId:
          type: string
          nullable: true
        customerEmail:
          type: string
          nullable: true
        providerReference:
          type: string
          nullable: true
        providerTransactionId:
          type: string
          nullable: true
        providerRefundReference:
          type: string
          nullable: true
        amountKobo:
          type: number
        reasonCode:
          type: string
        reasonText:
          type: string
          nullable: true
        status:
          type: string
          enum:
            - requested
            - approved
            - processing
            - succeeded
            - failed
            - cancelled
        failureReason:
          type: string
          nullable: true
        resolutionNote:
          type: string
          nullable: true
        requestedBy:
          type: string
          nullable: true
        resolvedBy:
          type: string
          nullable: true
        requestedAt:
          type: string
        processedAt:
          type: string
          nullable: true
        updatedAt:
          type: string
          nullable: true
      required:
        - id
        - paymentId
        - orderId
        - amountKobo
        - reasonCode
        - status
        - requestedAt
    AdminRefundListEnvelopeDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/AdminRefundRecordDto'
      required:
        - data
    AdminRefundEnvelopeDto:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/AdminRefundRecordDto'
      required:
        - data
...